Add logic to fetch instantly from clicked swatches (...)
The changeColor command is debounced by default as there is no "onChangeEnd" functionality within the color picker and hue slider components from Mantine. This means requests would otherwise be sent excessively as you move the color picker around.

This however is unnecessary for singly clicking color swatches, so rather than create a separate changeColor function with no debouncing function, we can instead set the debounce timer to 0 only when a swatch was clicked.

This is easily managed with a useRef boolean.